Now I have the airline, departure, and destination all parsed.
It was not fun.
I may decided to parse more information, but idk.
Oh yeah next commit I will add the functionality which will get the layover city, but that shouldnt be hard.  If it wasn't late I would just do it now -___-
Thanks for comming to my ted talk