wooohoooo the parser is 'complete'
I added functionality that allows it to parse all the flight routes using globs
I am thinking about maybe adding the coordinate location of each airport...

Ok now the next step is to create a map for all the airlines which will map each airline to a numerical key.  This is going to be key in the machine learning process since the airline string really means nothing to the computer.

I guess I will have to parse the time of the flight too.  For that I'll just use a regular expression to get rid of the 'h' and 'm'.  Then I'll convert the travel time to a floating point number eg: 1h 30m ----> 1.5