Cleaned everything up
One issue that isnt really an issue is that the machine has a hard time predicting nonstop flights between airline hubs.  e.g. an american airlines flight from CLT to LGA produces one stop in charlotte.... where the flight departs from.  Sure this can be fixed with storing the input and checking against the output, collecting more data, or training the machine differently.  More importantly for this version just understand that a layover in a departure or a destination city doesnt mean anything.... hopefully i will fix that later
but for now it doesnt 'really' matter